feat(www): support MTDA API from pytest
pytest requires sync calls: move async APIs from mtda.pytest to
mtda.asyncio.
