+++ mtda/assets/pytest.py
@@ -10,303 +10,116 @@
 # ---------------------------------------------------------------------------
 
 import asyncio
-import re
-import uuid
-
-from pyodide.http import pyfetch
-from urllib.parse import quote
+import mtda.asyncio
 
 # ---------------------------------------------------------------------------
 # Implement our Python APIs using REST endpoints
 # ---------------------------------------------------------------------------
 
 
-PY_SESSION_ID = uuid.uuid4()
+class Test:
+    @staticmethod
+    def init():
+        return mtda.asyncio.Test.init()
+
+    @staticmethod
+    def initialized():
+        return mtda.asyncio.Test.initialized()
+
+    @staticmethod
+    def setup():
+        assert Test.init() is True
+        return Console.clear()
+
+    @staticmethod
+    def sleep(seconds):
+        assert Test.initialized() is True
+        return Support.call(asyncio.sleep(seconds))
+
+    @staticmethod
+    def teardown():
+        assert Test.initialized() is True
+        return Console.clear()
 
 
 class Console:
     @staticmethod
-    async def clear():
-        """
-        Clear the consle ring buffer
-
-        Returns:
-            bool: True if the request was successful
-        """
-        response = await Support.get("/console-clear",
-                                     {"session": PY_SESSION_ID})
-        return response.ok
+    def clear():
+        return Support.call(mtda.asyncio.Console.clear())
 
     @staticmethod
-    async def dump():
-        """
-        Dump contents of the console ring-buffer
-
-        Returns:
-            str: contents of the ring buffer
-        """
-        response = await Support.get_data("/console-dump",
-                                          {"session": PY_SESSION_ID})
-        if 'result' in response:
-            return response['result']['content']
-        return None
+    def dump():
+        return Support.call(mtda.asyncio.Console.dump())
 
     @staticmethod
-    async def flush():
-        """
-        Flush contents of the console ring-buffer
-
-        Returns:
-            str: contents of the ring buffer before it was cleared
-        """
-        response = await Support.get_data("/console-flush",
-                                          {"session": PY_SESSION_ID})
-        if 'result' in response:
-            return response['result']['content']
-        return None
+    def flush():
+        return Support.call(mtda.asyncio.Console.flush())
 
     @staticmethod
-    async def head():
-        """
-        Pop the first line from the console ring-buffer
-
-        Returns:
-            str: first line of the console ring buffer
-        """
-        response = await Support.get_data("/console-head",
-                                          {"session": PY_SESSION_ID})
-        if 'result' in response:
-            return response['result']['content']
-        return None
+    def head():
+        return Support.call(mtda.asyncio.Console.head())
 
     @staticmethod
-    async def lines():
-        """
-        Get the number of lines available in console ring-buffer
-
-        Returns:
-            int: count of lines available in the console ring buffer
-        """
-        response = await Support.get_data("/console-lines",
-                                          {"session": PY_SESSION_ID})
-        if 'result' in response:
-            return int(response['result']['count'])
-        return None
+    def lines():
+        return Support.call(mtda.asyncio.Console.lines())
 
     @staticmethod
-    async def send(what):
-        """
-        Send text to the console
-
-        Parameters:
-            str: text to send
-
-        Returns:
-            bool: True if the request was successful
-        """
-        response = await Support.get("/console-send",
-                                     {"what": what,
-                                      "session": PY_SESSION_ID})
-        return response.ok
+    def send(what):
+        return Support.call(mtda.asyncio.Console.send(what))
 
     @staticmethod
-    async def tail():
-        """
-        Pop the last line from the console ring-buffer and remove others
-
-        Returns:
-            str: last line of the console ring buffer
-        """
-        response = await Support.get_data("/console-tail",
-                                          {"session": PY_SESSION_ID})
-        if 'result' in response:
-            return response['result']['content']
-        return None
+    def tail():
+        return Support.call(mtda.asyncio.Console.tail())
 
     @staticmethod
-    async def wait_for(what, errors=None, timeout=30, intervals=5,
-                       screen="", flush=True):
-        if isinstance(what, list) is False:
-            what = [what]
-
-        contents = screen
-        result = None
-
-        err_list = []
-        if errors:
-            if isinstance(errors, list) is False:
-                errors = [errors]
-            for expr in errors:
-                err_list.append(re.compile(expr))
-
-        exp_list = []
-        for expr in what:
-            exp_list.append(re.compile(expr))
-
-        while timeout > 0:
-            if flush is True:
-                contents += await Console.flush()
-            else:
-                contents = screen + await Console.dump()
-            # Check for errors
-            for expr in err_list:
-                if expr.search(contents):
-                    break
-            # Check for expected messages
-            for expr in exp_list:
-                if expr.search(contents):
-                    result = contents
-                    break
-            if result is not None:
-                break
-            await asyncio.sleep(intervals)
-            timeout -= intervals
-        return result
+    def wait_for(what, errors=None, timeout=30, intervals=5,
+                 screen="", flush=True):
+        return Support.call(mtda.asyncio.Console.wait_for(what, errors, timeout,
+                                                          intervals, screen,
+                                                          flush))
 
 
 class Storage:
-    async def commit():
-        """
-        Commit changes made to the shared storage to disk.
-        Note: shared storage shall be assigned to the HOST.
+    def commit():
+        return Support.call(mtda.asyncio.Storage.commit())
 
-        Returns:
-            bool: True if the request was successful
-        """
-        response = await Support.get("/storage-commit",
-                                     {"session": PY_SESSION_ID})
-        return response.ok
+    def rollback():
+        return Support.call(mtda.asyncio.Storage.rollback())
 
-    async def rollback():
-        """
-        Rollback changes made to the shared storage to disk.
-        Note: shared storage shall be assigned to the HOST.
+    def toggle():
+        return Support.call(mtda.asyncio.Storage.toggle())
 
-        Returns:
-            bool: True if the request was successful
-        """
-        response = await Support.get("/storage-rollback",
-                                     {"session": PY_SESSION_ID})
-        return response.ok
+    def to_host():
+        return Support.call(mtda.asyncio.Storage.to_host())
 
-    async def toggle():
-        """
-        Toggle the shared storage between the HOST and TARGET
-
-        Returns:
-            str: the new state ("HOST", "TARGET" or "???")
-        """
-        response = await Support.get_data("/storage-toggle",
-                                          {"session": PY_SESSION_ID})
-        if 'result' in response:
-            return response['result']['status']
-        return '???'
-
-    async def to_host():
-        """
-        Move shared storage to the HOST
-
-        Returns:
-            bool: True if the request was successful
-        """
-        response = await Support.get("/storage-to-host",
-                                     {"session": PY_SESSION_ID})
-        return response.ok
-
-    async def to_target():
-        """
-        Move shared storage to the TARGET
-
-        Returns:
-            bool: True if the request was successful
-        """
-        response = await Support.get("/storage-to-target",
-                                     {"session": PY_SESSION_ID})
-        return response.ok
+    def to_target():
+        return Support.call(mtda.asyncio.Storage.to_target())
 
 
 class Support:
     @staticmethod
-    def sanitize_params(params):
-        return {k: quote(str(v), safe='') for k, v in params.items()}
-
-    @staticmethod
-    async def get(endpoint, params=None):
-        if params:
-            safe_params = Support.sanitize_params(params)
-            query = "&".join(f"{k}={v}" for k, v in safe_params.items())
-            endpoint = f"{endpoint}?{query}"
-        return await pyfetch(endpoint, method="GET")
-
-    @staticmethod
-    async def get_data(endpoint, params=None):
-        response = await Support.get(endpoint, params)
-        return await response.json()
+    def call(fn):
+        loop = asyncio.get_event_loop()
+        return loop.run_until_complete(fn)
 
 
 class Target:
     @staticmethod
-    async def off():
-        """
-        Power OFF the target device
-
-        Returns:
-            bool: True if the request was successful
-        """
-        response = await Support.get("/target-off",
-                                     {"session": PY_SESSION_ID})
-        return response.ok
+    def off():
+        return Support.call(mtda.asyncio.Target.off())
 
     @staticmethod
-    async def on():
-        """
-        Power ON the target device
-
-        Returns:
-            bool: True if the request was successful
-        """
-        response = await Support.get("/target-on",
-                                     {"session": PY_SESSION_ID})
-        return response.ok
+    def on():
+        return Support.call(mtda.asyncio.Target.on())
 
     @staticmethod
-    async def status():
-        """
-        Query status of the target devide ("ON" or "OFF")
-
-        Returns:
-            str: the current state ("ON", "OFF" or "???")
-        """
-        response = await Support.get_data("/target-status",
-                                          {"session": PY_SESSION_ID})
-        if 'result' in response:
-            return response['result']['status']
-        return '???'
+    def status():
+        return Support.call(mtda.asyncio.Target.status())
 
     @staticmethod
-    async def toggle():
-        """
-        Toggle the power state of the target.
-
-        Returns:
-            str: the new state ("ON", "OFF" or "???")
-        """
-        response = await Support.get_data("/power-toggle",
-                                          {"session": PY_SESSION_ID})
-        if 'result' in response:
-            return response['result']['status']
-        return '???'
+    def toggle():
+        return Support.call(mtda.asyncio.Target.toggle())
 
     @staticmethod
-    async def uptime():
-        """
-        Query uptime of the target devide (in seconds)
-
-        Returns:
-            float: seconds since the target device was powered ON
-        """
-        response = await Support.get_data("/target-uptime",
-                                          {"session": PY_SESSION_ID})
-        if 'result' in response:
-            return float(response['result']['uptime'])
-        return None
+    def uptime():
+        return Support.call(mtda.asyncio.Target.uptime())
